Refactor explorer config into useKnockoutExplorer hook
