+++ src/Main.tsx
@@ -53,214 +53,22 @@ import "object.entries/auto";
 import "./Libs/is-integer-polyfill";
 import "url-polyfill/url-polyfill.min";
 
-initializeIcons();
-
-import { AuthType } from "./AuthType";
-
 import { initializeIcons } from "office-ui-fabric-react/lib/Icons";
-import { applyExplorerBindings } from "./applyExplorerBindings";
-import { configContext, initializeConfiguration, Platform } from "./ConfigContext";
-import Explorer from "./Explorer/Explorer";
-import React, { useEffect } from "react";
+import React from "react";
 import ReactDOM from "react-dom";
 import copyImage from "../images/Copy.svg";
 import hdeConnectImage from "../images/HdeConnectCosmosDB.svg";
 import refreshImg from "../images/refresh-cosmos.svg";
 import arrowLeftImg from "../images/imgarrowlefticon.svg";
 import { KOCommentEnd, KOCommentIfStart } from "./koComment";
-import { updateUserContext } from "./UserContext";
-import { CollectionCreation } from "./Shared/Constants";
-import { extractFeatures } from "./Platform/Hosted/extractFeatures";
-import { emulatorAccount } from "./Platform/Emulator/emulatorAccount";
-import { HostedExplorerChildFrame } from "./HostedExplorerChildFrame";
-import {
-  getDatabaseAccountKindFromExperience,
-  getDatabaseAccountPropertiesFromMetadata,
-} from "./Platform/Hosted/HostedUtils";
-import { DefaultExperienceUtility } from "./Shared/DefaultExperienceUtility";
-import { parseResourceTokenConnectionString } from "./Platform/Hosted/Helpers/ResourceTokenUtils";
-import { AccountKind, DefaultAccountExperience, ServerIds } from "./Common/Constants";
-import { listKeys } from "./Utils/arm/generatedClients/2020-04-01/databaseAccounts";
-import { SelfServeType } from "./SelfServe/SelfServeUtils";
+import { useConfig } from "./hooks/useConfig";
+import { useKnockoutExplorer } from "./hooks/useKnockoutExplorer";
+
+initializeIcons();
 
 const App: React.FunctionComponent = () => {
-  useEffect(() => {
-    initializeConfiguration().then(async (config) => {
-      let explorer: Explorer;
-      if (config.platform === Platform.Hosted) {
-        const win = (window as unknown) as HostedExplorerChildFrame;
-        explorer = new Explorer();
-        explorer.selfServeType(SelfServeType.none);
-        if (win.hostedConfig.authType === AuthType.EncryptedToken) {
-          // TODO: Remove window.authType
-          window.authType = AuthType.EncryptedToken;
-          // Impossible to tell if this is a try cosmos sub using an encrypted token
-          explorer.isTryCosmosDBSubscription(false);
-          updateUserContext({
-            accessToken: encodeURIComponent(win.hostedConfig.encryptedToken),
-          });
-
-          const apiExperience: string = DefaultExperienceUtility.getDefaultExperienceFromApiKind(
-            win.hostedConfig.encryptedTokenMetadata.apiKind
-          );
-          explorer.initDataExplorerWithFrameInputs({
-            databaseAccount: {
-              id: "",
-              // id: Main._databaseAccountId,
-              name: win.hostedConfig.encryptedTokenMetadata.accountName,
-              kind: getDatabaseAccountKindFromExperience(apiExperience),
-              properties: getDatabaseAccountPropertiesFromMetadata(win.hostedConfig.encryptedTokenMetadata),
-              tags: [],
-            },
-            subscriptionId: undefined,
-            resourceGroup: undefined,
-            masterKey: undefined,
-            hasWriteAccess: true, // TODO: we should embed this information in the token ideally
-            authorizationToken: undefined,
-            features: extractFeatures(),
-            csmEndpoint: undefined,
-            dnsSuffix: undefined,
-            serverId: ServerIds.productionPortal,
-            extensionEndpoint: configContext.BACKEND_ENDPOINT,
-            subscriptionType: CollectionCreation.DefaultSubscriptionType,
-            quotaId: undefined,
-            addCollectionDefaultFlight: explorer.flight(),
-            isTryCosmosDBSubscription: explorer.isTryCosmosDBSubscription(),
-          });
-          explorer.isAccountReady(true);
-        } else if (win.hostedConfig.authType === AuthType.ResourceToken) {
-          window.authType = AuthType.ResourceToken;
-          // Resource tokens can only be used with SQL API
-          const apiExperience: string = DefaultAccountExperience.DocumentDB;
-          const parsedResourceToken = parseResourceTokenConnectionString(win.hostedConfig.resourceToken);
-          updateUserContext({
-            resourceToken: parsedResourceToken.resourceToken,
-            endpoint: parsedResourceToken.accountEndpoint,
-          });
-          explorer.resourceTokenDatabaseId(parsedResourceToken.databaseId);
-          explorer.resourceTokenCollectionId(parsedResourceToken.collectionId);
-          if (parsedResourceToken.partitionKey) {
-            explorer.resourceTokenPartitionKey(parsedResourceToken.partitionKey);
-          }
-          explorer.initDataExplorerWithFrameInputs({
-            databaseAccount: {
-              id: "",
-              name: parsedResourceToken.accountEndpoint,
-              kind: AccountKind.GlobalDocumentDB,
-              properties: { documentEndpoint: parsedResourceToken.accountEndpoint },
-              tags: { defaultExperience: apiExperience },
-            },
-            subscriptionId: undefined,
-            resourceGroup: undefined,
-            masterKey: undefined,
-            hasWriteAccess: true, // TODO: we should embed this information in the token ideally
-            authorizationToken: undefined,
-            features: extractFeatures(),
-            csmEndpoint: undefined,
-            dnsSuffix: undefined,
-            serverId: ServerIds.productionPortal,
-            extensionEndpoint: configContext.BACKEND_ENDPOINT,
-            subscriptionType: CollectionCreation.DefaultSubscriptionType,
-            quotaId: undefined,
-            addCollectionDefaultFlight: explorer.flight(),
-            isTryCosmosDBSubscription: explorer.isTryCosmosDBSubscription(),
-            isAuthWithresourceToken: true,
-          });
-          explorer.isAccountReady(true);
-          explorer.isRefreshingExplorer(false);
-        } else if (win.hostedConfig.authType === AuthType.ConnectionString) {
-          // For legacy reasons lots of code expects a connection string login to look and act like an encrypted token login
-          window.authType = AuthType.EncryptedToken;
-          // Impossible to tell if this is a try cosmos sub using an encrypted token
-          explorer.isTryCosmosDBSubscription(false);
-          updateUserContext({
-            accessToken: encodeURIComponent(win.hostedConfig.encryptedToken),
-          });
-
-          const apiExperience: string = DefaultExperienceUtility.getDefaultExperienceFromApiKind(
-            win.hostedConfig.encryptedTokenMetadata.apiKind
-          );
-          explorer.initDataExplorerWithFrameInputs({
-            databaseAccount: {
-              id: "",
-              // id: Main._databaseAccountId,
-              name: win.hostedConfig.encryptedTokenMetadata.accountName,
-              kind: getDatabaseAccountKindFromExperience(apiExperience),
-              properties: getDatabaseAccountPropertiesFromMetadata(win.hostedConfig.encryptedTokenMetadata),
-              tags: [],
-            },
-            subscriptionId: undefined,
-            resourceGroup: undefined,
-            masterKey: win.hostedConfig.masterKey,
-            hasWriteAccess: true, // TODO: we should embed this information in the token ideally
-            authorizationToken: undefined,
-            features: extractFeatures(),
-            csmEndpoint: undefined,
-            dnsSuffix: undefined,
-            serverId: ServerIds.productionPortal,
-            extensionEndpoint: configContext.BACKEND_ENDPOINT,
-            subscriptionType: CollectionCreation.DefaultSubscriptionType,
-            quotaId: undefined,
-            addCollectionDefaultFlight: explorer.flight(),
-            isTryCosmosDBSubscription: explorer.isTryCosmosDBSubscription(),
-          });
-          explorer.isAccountReady(true);
-        } else if (win.hostedConfig.authType === AuthType.AAD) {
-          window.authType = AuthType.AAD;
-          const account = win.hostedConfig.databaseAccount;
-          const accountResourceId = account.id;
-          const subscriptionId = accountResourceId && accountResourceId.split("subscriptions/")[1].split("/")[0];
-          const resourceGroup = accountResourceId && accountResourceId.split("resourceGroups/")[1].split("/")[0];
-          updateUserContext({
-            authorizationToken: `Bearer ${win.hostedConfig.authorizationToken}`,
-            databaseAccount: win.hostedConfig.databaseAccount,
-          });
-          const keys = await listKeys(subscriptionId, resourceGroup, account.name);
-          explorer.initDataExplorerWithFrameInputs({
-            databaseAccount: account,
-            subscriptionId,
-            resourceGroup,
-            masterKey: keys.primaryMasterKey,
-            hasWriteAccess: true, //TODO: 425017 - support read access
-            authorizationToken: `Bearer ${win.hostedConfig.authorizationToken}`,
-            features: extractFeatures(),
-            csmEndpoint: undefined,
-            dnsSuffix: undefined,
-            serverId: ServerIds.productionPortal,
-            extensionEndpoint: configContext.BACKEND_ENDPOINT,
-            subscriptionType: CollectionCreation.DefaultSubscriptionType,
-            quotaId: undefined,
-            addCollectionDefaultFlight: explorer.flight(),
-            isTryCosmosDBSubscription: explorer.isTryCosmosDBSubscription(),
-          });
-          explorer.isAccountReady(true);
-        }
-      } else if (config.platform === Platform.Emulator) {
-        window.authType = AuthType.MasterKey;
-        explorer = new Explorer();
-        explorer.selfServeType(SelfServeType.none);
-        explorer.databaseAccount(emulatorAccount);
-        explorer.isAccountReady(true);
-      } else if (config.platform === Platform.Portal) {
-        explorer = new Explorer();
-
-        // In development mode, try to load the iframe message from session storage.
-        // This allows webpack hot reload to funciton properly
-        if (process.env.NODE_ENV === "development") {
-          const initMessage = sessionStorage.getItem("portalDataExplorerInitMessage");
-          if (initMessage) {
-            const message = JSON.parse(initMessage);
-            console.warn("Loaded cached portal iframe message from session storage");
-            console.dir(message);
-            explorer.initDataExplorerWithFrameInputs(message);
-          }
-        }
-
-        window.addEventListener("message", explorer.handleMessage.bind(explorer), false);
-      }
-      applyExplorerBindings(explorer);
-    });
-  }, []);
+  const config = useConfig();
+  useKnockoutExplorer(config);
 
   return (
     <div className="flexContainer">
